addressing comments; hash_to_curve -> hash_to_group
1
